Move auth package.
