+++ internal/codeintel/uploads/transport/http/auth/gitlab.go
