+++ internal/codeintel/uploads/transport/http/auth/gitlab_test.go
