+++ internal/codeintel/uploads/transport/http/auth/iface.go
