+++ internal/codeintel/uploads/transport/http/auth/util.go
