+++ mockgen.test.yaml
@@ -41,8 +41,8 @@
   path: github.com/sourcegraph/sourcegraph/internal/uploadhandler
   interfaces:
     - DBStore
-- filename: internal/codeintel/uploads/transport/httpapi/auth/mocks_test.go
-  path: github.com/sourcegraph/sourcegraph/internal/codeintel/uploads/transport/httpapi/auth
+- filename: internal/codeintel/uploads/transport/http/auth/mocks_test.go
+  path: github.com/sourcegraph/sourcegraph/internal/codeintel/uploads/transport/http/auth
   interfaces:
     - GitHubClient
 - filename: enterprise/cmd/frontend/internal/codeintel/resolvers/mocks_test.go
